refactor: move RegisterForm into a separate component for better reusability
