+++ frontend/src/app/components/features/registerForm/RegisterForm.module.scss
@@ -1,0 +1,85 @@
+@import "../../../variables.module.scss";
+
+.buttonWrapper {
+  display: flex;
+  justify-content: center;
+}
+
+.icon {
+  width: 1.2rem;
+  height: 1.2rem;
+}
+
+.form {
+  width: 100%;
+}
+
+.required {
+  color: $red-alert;
+  font-size: 0.8rem;
+  margin-bottom: 0;
+}
+
+.selectWrapper {
+  display: flex;
+  align-items: center;
+  position: relative;
+  border: 1px solid #ccc;
+  border-radius: 0.3rem;
+  padding: 0.5rem;
+  margin: 0.25rem 0 1rem;
+
+  &:focus-within {
+    border-color: $blue_primary;
+  }
+}
+
+.select {
+  width: 100%;
+  border: none;
+  padding-left: 0.3rem;
+
+  &:focus {
+    outline: none;
+  }
+}
+
+.label {
+  font-size: 0.9rem;
+  margin-bottom: 5rem;
+
+  input {
+    margin-right: 0.25rem;
+  }
+}
+
+.privacyPolicy {
+  width: 100%;
+  padding: 0.3rem;
+  border: 1px solid #ccc;
+  border-radius: 0.3rem;
+  font-size: 0.8rem;
+  max-height: 60px;
+  overflow-y: auto;
+
+  margin: 0.25rem 0;
+}
+
+.checkboxWrapper {
+  margin-bottom: 1.5rem;
+}
+
+.icon {
+  width: 1.2rem;
+  height: 1.2rem;
+}
+
+.errorText {
+  color: $red-alert;
+  font-size: 0.8rem;
+
+  &--prefecture {
+    margin-top: -0.9rem;
+    margin-bottom: 1rem;
+  }
+}
