+++ frontend/src/app/customers/register/page.tsx
@@ -1,128 +1,14 @@
 "use client";
 
-import { useInput } from "@/app/hooks/useInput";
 import { useRouter } from "next/navigation";
-import React, { FormEvent, useEffect, useState } from "react";
-import { prefectures } from "@/app/helper/data/data";
+import React from "react";
 import styles from "./page.module.scss";
 import Image from "next/image";
 import Link from "next/link";
-import TextInput from "@/app/components/elements/textInput/TextInput";
-import {
-  EnvelopeIcon,
-  HomeIcon,
-  LockClosedIcon,
-  UserCircleIcon,
-} from "@heroicons/react/24/outline";
-import ActionButton from "@/app/components/elements/buttons/actionButton/ActionButton";
-import { toast } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
-import { useSelect } from "@/app/hooks/useSelect";
-import { registerCustomer } from "@/app/helper/api/customersApi";
-import { customerRegisterSchema } from "@/app/schemas/customerRegisterSchema";
-import zxcvbn from "zxcvbn";
-import PasswordStrengthMeter from "@/app/components/elements/passwordStrengthMeter/PasswordStrengthMeter";
+import RegisterForm from "@/app/components/features/registerForm/RegisterForm";
 
 function Register() {
-  const [name, onNameChange] = useInput();
-  const [email, onEmailChange] = useInput();
-  const [password, onPasswordChange] = useInput();
-  const [passConfirmation, onPassConfirmationChange] = useInput();
-  const [prefecture, setPrefecture, onPrefectureChange] = useSelect("");
-  const [isAgreed, setIsAgreed] = useState(false);
-  const [errors, setErrors] = useState<Record<string, string>>({});
-  const [passwordStrength, setPasswordStrength] = useState<number>(0);
-  const [passwordFeedback, setPasswordFeedback] = useState<string>("");
-  const [showPassword, setShowPassword] = useState(false);
-
   const router = useRouter();
-
-  // Real-time password strength check
-  useEffect(() => {
-    if (password) {
-      setErrors((prev) => ({ ...prev, password: "" }));
-      const result = zxcvbn(password);
-      setPasswordStrength(result.score);
-      setPasswordFeedback(result.feedback.suggestions.join(" "));
-    } else {
-      setPasswordStrength(0);
-      setPasswordFeedback("");
-    }
-  }, [password]);
-
-  const togglePasswordVisibility = () => {
-    setShowPassword((prevState) => !prevState);
-  };
-
-  const registerHandler = async (e: FormEvent) => {
-    e.preventDefault();
-
-    setErrors({});
-
-    // Validate form using Zod
-    const validationResult = customerRegisterSchema.safeParse({
-      name,
-      email,
-      password,
-      passConfirmation,
-      prefecture,
-      isAgreed,
-    });
-
-    // If validation fails, extract error messages
-    if (!validationResult.success) {
-      const validationErrors: Record<string, string> = {};
-
-      validationResult.error.errors.forEach((err) => {
-        if (err.path[0]) {
-          validationErrors[err.path[0]] = err.message;
-        }
-      });
-
-      setErrors(validationErrors);
-      return;
-    }
-
-    // Prevent submission if password is weak
-    if (passwordStrength < 3) {
-      setErrors({
-        password:
-          "Your password is too weak. Try using a longer passphrase or a password manager.",
-      });
-
-      // Refocus password field for browser suggestions
-      document.getElementById("password")?.focus();
-      return;
-    }
-
-    try {
-      const response = await registerCustomer({
-        name,
-        email,
-        password,
-        prefecture,
-      });
-
-      if (response.status === 409) {
-        setErrors({ email: response.message });
-        return;
-      }
-
-      const successMessage = response.message || "Registration successful!";
-      const redirectUrl = response.redirectUrl || "/customers/login";
-
-      toast.success(successMessage);
-
-      router.push(redirectUrl);
-    } catch (error) {
-      setErrors({
-        unexpectedError:
-          error instanceof Error
-            ? error.message
-            : "An unexpected error occurred.",
-      });
-    }
-  };
 
   return (
     <main>
@@ -141,134 +27,7 @@ function Register() {
             Already a member? <Link href="/customers/login">Log In</Link>
           </p>
 
-          <form className={styles.form} onSubmit={registerHandler}>
-            <p className={styles.required}>*Required</p>
-            <TextInput
-              label="Name"
-              type="text"
-              value={name}
-              placeholder="e.g., John Doe"
-              onChange={onNameChange}
-              icon={<UserCircleIcon className={styles.icon} />}
-              inputRequired={true}
-              autoComplete="name"
-              error={errors.name}
-            />
-
-            <TextInput
-              label="Email"
-              type="email"
-              value={email}
-              placeholder="e.g., example@aaasobo.com"
-              onChange={onEmailChange}
-              icon={<EnvelopeIcon className={styles.icon} />}
-              inputRequired={true}
-              autoComplete="email"
-              error={errors.email}
-            />
-
-            <TextInput
-              id="password"
-              label="Password"
-              type="password"
-              value={password}
-              placeholder="Create a password (8+ characters)"
-              onChange={onPasswordChange}
-              icon={<LockClosedIcon className={styles.icon} />}
-              inputRequired={true}
-              minLength={8}
-              autoComplete="new-password"
-              error={errors.password}
-              showPassword={showPassword}
-              onTogglePasswordVisibility={togglePasswordVisibility}
-            />
-
-            <PasswordStrengthMeter
-              password={password}
-              passwordStrength={passwordStrength}
-              passwordFeedback={passwordFeedback}
-            />
-
-            <TextInput
-              label="Password Confirmation"
-              type="password"
-              value={passConfirmation}
-              placeholder="Re-enter your password"
-              onChange={onPassConfirmationChange}
-              icon={<LockClosedIcon className={styles.icon} />}
-              inputRequired={true}
-              autoComplete="new-password"
-              error={errors.passConfirmation}
-              showPassword={showPassword}
-              onTogglePasswordVisibility={togglePasswordVisibility}
-            />
-
-            <label className={styles.label}>
-              Prefecture of Residence
-              <span className={styles.required}>*</span>
-              <div className={styles.selectWrapper}>
-                <HomeIcon className={styles.icon} />
-                <select
-                  className={styles.select}
-                  value={prefecture}
-                  onChange={onPrefectureChange}
-                  required
-                  style={{ color: prefecture ? "black" : "gray" }}
-                >
-                  <option value="" disabled>
-                    Select your prefecture of residence
-                  </option>
-                  {prefectures.map((prefecture) => (
-                    <option key={prefecture} value={prefecture}>
-                      {prefecture}
-                    </option>
-                  ))}
-                </select>
-              </div>
-              {errors.prefecture && (
-                <p
-                  className={`${styles.errorText} ${styles["errorText--prefecture"]}`}
-                >
-                  {errors.prefecture}
-                </p>
-              )}
-            </label>
-
-            <label className={styles.label}>
-              Privacy Policy
-              <span className={styles.required}>*</span>
-              <p className={styles.privacyPolicy}>
-                We will take a screenshot as a record that the class was
-                conducted. Additionally, we may record the session for the
-                purpose of improving the instructor&apos;s skills.
-              </p>
-            </label>
-            <div className={styles.checkboxWrapper}>
-              <label className={styles.label}>
-                <input
-                  type="checkbox"
-                  checked={isAgreed}
-                  onChange={() => setIsAgreed(!isAgreed)}
-                  required
-                />
-                I agree.
-              </label>
-              {errors.isAgreed && (
-                <p className={styles.errorText}>{errors.isAgreed}</p>
-              )}
-            </div>
-
-            <div className={styles.buttonWrapper}>
-              <ActionButton
-                btnText="Create Account"
-                className="bookBtn"
-                type="submit"
-              />
-            </div>
-            {errors.unexpectedError && (
-              <p className={styles.errorText}>{errors.unexpectedError}</p>
-            )}
-          </form>
+          <RegisterForm userType="customer" onSuccessRedirect={router.push} />
         </div>
       </div>
     </main>
