+++ frontend/src/app/schemas/customerAuthSchema.ts
